Added registration and login
You can now register an account and login to that account, however I could not figure out how to get edit profile to work, passwords also do not have encryption yet. Lastly I made some changes to the registration form, requiring the user to input their email address and password twice before being able to submit, currently there is no actual email being sent to the email they provided, lastly, the checkbox for "register for promotion" is off center, so if someone can change that so it's flush with the other fields in the sign up page that would be great.